new: Modularizes core connection and prompt logic
Splits connection workflow, resource selection, inference, and user prompts into modular components to improve code maintainability and extensibility. Introduces filterable, user-friendly selection UIs and clearer connection/recommendation summaries. Enhances CLI usability by supporting zoxide-style real-time search and filter hints, paving the way for easier customization and future feature expansion.
