+++ src/aws-port-forward-new.ts
@@ -1,0 +1,5 @@
+// Re-export the main connection functions from the modular implementation
+export {
+  connectToRDS,
+  connectToRDSWithInference,
+} from "./core/connection-flow.js";
